mlrho version 1.24 commit
Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   Makefile
	modified:   deltaComp.c
	modified:   eprintf.h
	modified:   interface.c
	modified:   interface.h
	modified:   mlComp.c
	modified:   mlComp.h
	modified:   mlRho.c
	modified:   piComp.c
	modified:   profileTree.c
	new file:   queue.h
	modified:   valgrind.sh

+++ profileTree.c
@@ -18,7 +18,7 @@
 #include "profileTree.h"
 #include "ld.h"
 
-int numPos;
+double numPos;
 Node **resetProfilePairs(Node **profilePairs, int numProfiles);
 Node **countPairs(Node **pairs, int numProfiles, ContigDescr *contigDescr, FILE *fp, int dist);
